+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Additional logs to indicate save/restore completion.
 - Support for None leaves in PyTree save/restore.
+- `read: bool` option on all_steps to force read from storage location instead
+of using cached steps.
 
 ### Fixed
 
-- Usage of bytes_limiter to prevent too many bytes from being read during a single restore call.
+- Usage of bytes_limiter to prevent too many bytes from being read during a
+single restore call.
 
 ## [0.0.23] - 2022-12-08